pyproject.toml
[tool.poetry]
name = "django-shorts"
version = "0.1.0"
description = ""
authors = ["temapskov <psk@ro.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
django-ninja = "^1.3.0"
pillow = "^11.0.0"
